warning fix
git-svn-id: https://stlport.svn.sourceforge.net/svnroot/stlport/branches/STLPORT_5_0/STLport@1227 01b27279-081b-0410-8cf9-d9d9e0fb0389

+++ test/unit/unordered_test.cpp
@@ -159,7 +159,7 @@ void UnorderedTest::umap()
     }
   }
 
-  CPPUNIT_ASSERT( us.size() == NB_ELEMS );
+  CPPUNIT_ASSERT( us.size() == (size_t)NB_ELEMS );
   vector<pair<int, int> > us_val;
 
   umaptype::local_iterator lit, litEnd;
